Add PedestalParamHandler
PedestalModeHandler sends pedestal frames, loops, and the enabled state in one command.  This means the values are only updated on a mode change.  This new handler calls PedestalModeHandler with its current state whenever one of the pedestal params is changed, ensuring the values can be updated even when the mode is enabled